# api-server/api/oas-doc.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Contacts API
  description: An API to manage contacts. Generated with OAS TOOLS CLI 

paths:
  /api/v1/contacts:
    get:
      description: Returns all contacts
      operationId: getContact
      responses:
        200:
          description: Contact list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/contact'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      description: Creates a new contact
      operationId: addContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contact'
        description: Contact to be created
        required: true
      responses:
        201:
          description: Contact
        default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/error'
  /api/v1/contacts/{name}:
    get:
      description: Returns a single contact by its name
      operationId: findByname
      parameters:
        - required: true
          name: name
          description: name of the contact to fetch
          in: path
          schema:
            type: string
      responses:
        200:
          description: Contact found by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contact'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      description: Update a contact based on its name
      operationId: updateContact
      parameters:
        - required: true
          name: name
          description: name of the contact to update
          in: path
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contact'
        description: Contact to be updated
        required: true
      responses:
        '204':
          description: Contact updated
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      description: Deletes a single contact based on its name
      operationId: deleteContact
      parameters:
        - required: true
          name: name
          description: name of the contact to delete
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Contact deleted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

components:
  schemas:
    contact:
        type: object
        required:
          - name
        properties:
          name:
            type: string
          email:
            type: string
          phone:
            type: number
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      additionalProperties: true